Update Northwestern (PHU ON).
Remove tables with 0 rows before further processing. There is a "ghost" version of the outcomes table that comes before the real one with the fatality numbers. This is probably a result of adding the JS parameter to the dataset, such that the package now uses Selenium to open the dataset rather than a simple request.